Optimized SharedList for Native
